+++ packages/plugin-ext/src/hosted/node/scanners/grammars-reader.ts
@@ -22,10 +22,10 @@ import * as fs from '@theia/core/shared/fs-extra';
 @injectable()
 export class GrammarsReader {
 
-    readGrammars(rawGrammars: PluginPackageGrammarsContribution[], pluginPath: string): GrammarsContribution[] {
+    async readGrammars(rawGrammars: PluginPackageGrammarsContribution[], pluginPath: string): Promise<GrammarsContribution[]> {
         const result = new Array<GrammarsContribution>();
         for (const rawGrammar of rawGrammars) {
-            const grammar = this.readGrammar(rawGrammar, pluginPath);
+            const grammar = await this.readGrammar(rawGrammar, pluginPath);
             if (grammar) {
                 result.push(grammar);
             }
@@ -34,13 +34,14 @@ export class GrammarsReader {
         return result;
     }
 
-    private readGrammar(rawGrammar: PluginPackageGrammarsContribution, pluginPath: string): GrammarsContribution | undefined {
+    private async readGrammar(rawGrammar: PluginPackageGrammarsContribution, pluginPath: string): Promise<GrammarsContribution | undefined> {
         // TODO: validate inputs
         let grammar: string | object;
+
         if (rawGrammar.path.endsWith('json')) {
-            grammar = fs.readJSONSync(path.resolve(pluginPath, rawGrammar.path));
+            grammar = await fs.readJSON(path.resolve(pluginPath, rawGrammar.path));
         } else {
-            grammar = fs.readFileSync(path.resolve(pluginPath, rawGrammar.path), 'utf8');
+            grammar = await fs.readFile(path.resolve(pluginPath, rawGrammar.path), 'utf8');
         }
         return {
             language: rawGrammar.language,